move masks to new package
