+++ packages/masks/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@googleforcreators/masks",
+  "private": true,
+  "engines": {
+    "node": ">= 16",
+    "npm": "~7.3"
+  },
+  "exports": {
+    "default": "./src/index.js"
+  },
+  "main": "./src/index.js",
+  "type": "module",
+  "dependencies": {
+    "@googleforcreators/elements":"*",
+    "@googleforcreators/fonts": "*",
+    "@googleforcreators/i18n": "*",
+    "@googleforcreators/media":"*",
+    "@googleforcreators/patterns": "*",
+    "@googleforcreators/react": "*",
+    "@googleforcreators/transform": "*",
+    "@googleforcreators/units": "*",
+    "react": "^16.14.0",
+    "react-dom": "^16.14.0",
+    "styled-components": "^5.2.1",
+    "uuid": "^8.3.2"
+  },
+  "devDependencies": {
+    "prop-types": "^15.8.0",
+    "@testing-library/react": "^11.2.3"
+  }
+}
