--- packages/media/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@googleforcreators/media",
-  "description": "Functionality for creating and working with media resources in the Web Stories editor.",
-  "private": false,
-  "version": "0.0.1",
-  "author": {
-    "name": "Google",
-    "url": "https://creators.google/"
-  },
-  "license": "Apache-2.0",
-  "keywords": [
-    "web stories"
-  ],
-  "homepage": "https://github.com/googleforcreators/web-stories-wp/",
-  "repository": "git+https://github.com/GoogleForCreators/web-stories-wp.git",
-  "bugs": {
-    "url": "https://github.com/googleforcreators/web-stories-wp/issues"
-  },
-  "engines": {
-    "node": ">= 16",
-    "npm": ">= 7.3"
-  },
-  "type": "module",
-  "exports": {
-    "default": "./src/index.js"
-  },
-  "main": "./src/index.js",
-  "sideEffects": false,
-  "dependencies": {
-    "@googleforcreators/react": "*",
-    "@googleforcreators/tracking": "*",
-    "@googleforcreators/units": "*",
-    "mime": "^3.0.0",
-    "prop-types": "^15.8.1"
-  },
-  "devDependencies": {}
-}
